feat: 편지 쓰기 개발 및 예외처리 일부 개발 (#17)
